Move application setup to base testcase
The Base TestCase is now responsible for creating the application object
while the Web TestCase handles all web related things.

If the refresh database trait is added to a class, the database will
be reset in between functions.

This eliminates the need for calling setup database and parent::setup
to get the functionality of both.
